+++ tests/functional/Extension/HeadingPermalink/HeadingPermalinkExtensionTest.php
@@ -31,7 +31,7 @@ final class HeadingPermalinkExtensionTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \trim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \trim((string) $converter->convertToHtml($input)));
     }
 
     public function dataProviderForTestHeadingPermalinksWithDefaultOptions(): \Generator
@@ -64,7 +64,7 @@ final class HeadingPermalinkExtensionTest extends TestCase
 
         $converter = new CommonMarkConverter($config, $environment);
 
-        $this->assertEquals($expected, \trim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \trim((string) $converter->convertToHtml($input)));
     }
 
     public function dataProviderForTestHeadingPermalinksWithCustomOptions(): \Generator
@@ -90,7 +90,7 @@ final class HeadingPermalinkExtensionTest extends TestCase
         $input    = '# Hello World!';
         $expected = \sprintf('<h1><a id="hello-world" href="#hello-world" name="hello-world" class="heading-permalink" aria-hidden="true" title="Permalink">%s</a>Hello World!</h1>', HeadingPermalinkRenderer::DEFAULT_SYMBOL);
 
-        $this->assertEquals($expected, \trim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \trim((string) $converter->convertToHtml($input)));
     }
 
     public function testHeadingPermalinksWithEmptySymbol(): void
@@ -109,7 +109,7 @@ final class HeadingPermalinkExtensionTest extends TestCase
         $input    = '# Hello World!';
         $expected = '<h1><a id="user-content-hello-world" href="#hello-world" name="hello-world" class="heading-permalink" aria-hidden="true" title="Permalink"></a>Hello World!</h1>';
 
-        $this->assertEquals($expected, \trim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \trim((string) $converter->convertToHtml($input)));
     }
 
     public function testHeadingPermalinksWithInvalidInsertConfigurationValue(): void
@@ -156,6 +156,6 @@ EOT;
 <h4>4</h4>
 EOT;
 
-        $this->assertEquals($expected, \trim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \trim((string) $converter->convertToHtml($input)));
     }
 }