+++ tests/functional/Extension/Mention/MentionParserTest.php
@@ -37,7 +37,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserWithoutSpaceInFront(): void
@@ -52,7 +52,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserWithNonMatchingSymbol(): void
@@ -67,7 +67,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserWithNonMatchingRegex(): void
@@ -82,7 +82,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserWithNullUrl(): void
@@ -100,7 +100,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserUsingCallback(): void
@@ -124,7 +124,7 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 
     public function testMentionParserUsingCallbackReturnsAbstractInline(): void
@@ -147,6 +147,6 @@ final class MentionParserTest extends TestCase
 
         $converter = new CommonMarkConverter([], $environment);
 
-        $this->assertEquals($expected, \rtrim($converter->convertToHtml($input)));
+        $this->assertEquals($expected, \rtrim((string) $converter->convertToHtml($input)));
     }
 }